requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
jsonschema==4.25.1
pyyaml==6.0.3
pytest==9.0.3
httpx==0.28.1
